compute: move Routine schema from plugin-routine (wiring stays as makeRoutine/wireTriggers)
