+++ packages/core/compute/compute/src/types/Trigger.ts
@@ -185,7 +185,7 @@ export class Trigger extends Type.makeObject<Trigger>(DXN.make('org.dxos.type.tr
   Schema.Struct({
     /**
      * Runnable (operation or workflow) to invoke.
-     * Wired programmatically (see `Routine.wireTriggers`); not user-editable, so hidden from forms.
+     * Wired programmatically (see plugin-routine's `wireTriggers`); not user-editable, so hidden from forms.
      */
     runnable: Ref.Ref(Runnable.Runnable).pipe(
       Schema.annotations({ title: 'Runnable' }),
