+++ packages/plugins/plugin-assistant/src/templates/scaffold.ts
@@ -4,9 +4,9 @@
 
 import * as Effect from 'effect/Effect';
 
-import { Instructions, Skill, Trigger } from '@dxos/compute';
+import { Instructions, type Routine, Skill, Trigger } from '@dxos/compute';
 import { Ref } from '@dxos/echo';
-import { Routine } from '@dxos/plugin-routine/types';
+import { makeRoutine } from '@dxos/plugin-routine';
 
 export type ScheduledRoutineOptions = {
   name: string;
@@ -30,7 +30,7 @@ export const makeScheduledRoutine = ({
 }: ScheduledRoutineOptions): Effect.Effect<Routine.Routine, never, never> => {
   const skills = skillKeys.map((key) => Ref.fromURI(Skill.registryURI(key)));
   return Effect.succeed(
-    Routine.make({
+    makeRoutine({
       name,
       instructions: Instructions.make({ name, text, skills }),
       trigger: Trigger.make({ spec: Trigger.specTimer(cron), enabled: false }),
