+++ packages/plugins/plugin-routine/src/capabilities/react-surface.tsx
@@ -8,11 +8,10 @@ import React from 'react';
 import { Capabilities, Capability } from '@dxos/app-framework';
 import { Surface } from '@dxos/app-framework/ui';
 import { AppSurface, useActiveSpace } from '@dxos/app-toolkit/ui';
-import { Skill } from '@dxos/compute';
+import { Routine, Skill } from '@dxos/compute';
 
 import { RoutineArticle, RoutineCompanion, RoutineSettings, RoutineTraceCompanion, SkillArticle } from '#containers';
 import { meta } from '#meta';
-import { Routine } from '#types';
 
 export default Capability.makeModule(() =>
   Effect.succeed(
