+++ packages/plugins/plugin-routine/src/components/RoutineForm/RoutineForm.tsx
@@ -5,15 +5,15 @@
 import * as Schema from 'effect/Schema';
 import React, { type PropsWithChildren, useCallback, useMemo } from 'react';
 
-import { Instructions, Operation, Trigger } from '@dxos/compute';
+import { Instructions, Operation, Routine, Trigger } from '@dxos/compute';
 import { type Database, DXN, Entity, Filter, Obj, Query, Ref, Scope, Type } from '@dxos/echo';
 import { useObject, useQuery } from '@dxos/echo-react';
 import { ToggleGroup, ToggleGroupItem, composable, composableProps, useTranslation } from '@dxos/react-ui';
 import { Form, type FormFieldMap, RefField } from '@dxos/react-ui-form';
 
 import { meta } from '#meta';
-import { Routine } from '#types';
 
+import { wireTriggers } from '../../util';
 import { InstructionsEditor } from '../InstructionsEditor';
 import { TriggerEditor } from '../TriggerEditor';
 
@@ -119,7 +119,7 @@ const Section = ({ title, children }: PropsWithChildren<{ title: string }>) => (
  * Single action: an Operation (the routine's `spec.runnable`) or an owned Instructions edited inline. The
  * action kind is derived from the routine's `spec` — an absent spec means "operation, none chosen yet" (a
  * scaffolded routine always carries an instructions spec, so the only way to clear it is switching to an
- * operation). `Routine.make` establishes the owned-instructions wiring when the routine is scaffolded.
+ * operation). `makeRoutine` establishes the owned-instructions wiring when the routine is scaffolded.
  */
 const ActionEditor = ({
   db,
@@ -147,7 +147,7 @@ const ActionEditor = ({
         routine.spec = operation ? { kind: 'runnable', runnable: operation } : undefined;
       });
       // Keep the owned trigger's `function`/`input` in sync with the new action.
-      Routine.wireTriggers(routineProp);
+      wireTriggers(routineProp);
     },
     [updateRoutine, routineProp],
   );
@@ -167,7 +167,7 @@ const ActionEditor = ({
         }
       });
       // Re-wire the owned trigger to dispatch the new action (RunInstructions vs the operation).
-      Routine.wireTriggers(routineProp);
+      wireTriggers(routineProp);
     },
     [updateRoutine, routineProp],
   );
