+++ packages/plugins/plugin-routine/src/components/TriggerEditor/TriggerEditor.stories.tsx
@@ -6,7 +6,7 @@ import { type Meta, type StoryObj } from '@storybook/react-vite';
 import React from 'react';
 import { expect, userEvent, within } from 'storybook/test';
 
-import { Trigger } from '@dxos/compute';
+import { Routine, Trigger } from '@dxos/compute';
 import { Feed, Filter, Obj } from '@dxos/echo';
 import { useQuery } from '@dxos/echo-react';
 import { type Space } from '@dxos/react-client/echo';
@@ -14,7 +14,6 @@ import { useClientStory, withClientProvider } from '@dxos/react-client/testing';
 import { Loading, withLayout, withTheme } from '@dxos/react-ui/testing';
 
 import { translations } from '#translations';
-import { Routine } from '#types';
 
 import { TriggerEditor } from './TriggerEditor';
