+++ packages/plugins/plugin-routine/src/components/TriggerEditor/TriggerEditor.tsx
@@ -5,7 +5,7 @@
 import * as Schema from 'effect/Schema';
 import React, { useCallback, useMemo, useState } from 'react';
 
-import { Trigger } from '@dxos/compute';
+import { Routine, Trigger } from '@dxos/compute';
 import { type Database, DXN, Feed, Filter, Obj, Query, Ref, Scope, Type } from '@dxos/echo';
 import { useQuery } from '@dxos/echo-react';
 import { IconButton, Input, ThemedClassName, useTranslation } from '@dxos/react-ui';
@@ -14,8 +14,8 @@ import { ParentLabelAnnotation } from '@dxos/schema';
 import { mx } from '@dxos/ui-theme';
 
 import { meta } from '#meta';
-import { Routine } from '#types';
 
+import { wireTriggers } from '../../util';
 import {
   FrequencyDefaults,
   Schedule,
@@ -342,7 +342,7 @@ const useTriggerForm = (routine: Routine.Routine, trigger?: Trigger.Trigger) => 
       const remote = values.remote;
       // Edit the spec, `enabled`, and `remote` on the trigger directly from the form values.
       // The trigger's `function` and `input` (including the instructions binding and any operation-specific
-      // bindings like `{ magazine }`) are wired once by `Routine.make`, so they are not re-derived here.
+      // bindings like `{ magazine }`) are wired once by `makeRoutine`, so they are not re-derived here.
       if (trigger) {
         Obj.update(trigger, (trigger) => {
           // The subscription spec's QueryAST is deeply readonly while the live ECHO draft's `spec` is mutable;
@@ -353,7 +353,7 @@ const useTriggerForm = (routine: Routine.Routine, trigger?: Trigger.Trigger) => 
           trigger.remote = remote;
         });
       } else {
-        // Defensive: the draft normally carries an owned trigger already (see `Routine.make`). If absent,
+        // Defensive: the draft normally carries an owned trigger already (see `makeRoutine`). If absent,
         // create one in memory and attach it to the routine graph — nothing is persisted until save.
         const created = Trigger.make({ spec, enabled, remote });
         Obj.setParent(created, routine);
@@ -362,7 +362,7 @@ const useTriggerForm = (routine: Routine.Routine, trigger?: Trigger.Trigger) => 
         });
         // Wire the new trigger's `function`/`input` to dispatch the routine's action (RunInstructions binds the
         // owned instructions; an operation binds directly).
-        Routine.wireTriggers(routine);
+        wireTriggers(routine);
       }
     },
     [routine, trigger],
