+++ packages/plugins/plugin-routine/src/containers/RoutineCompanion/RoutineCompanion.tsx
@@ -8,6 +8,7 @@ import React, { useCallback, useMemo, useRef, useState } from 'react';
 
 import { useCapabilities, useOperationInvoker } from '@dxos/app-framework/ui';
 import { type AppSurface } from '@dxos/app-toolkit/ui';
+import { Routine } from '@dxos/compute';
 import { Database, Filter, Obj, Type } from '@dxos/echo';
 import { useQuery } from '@dxos/echo-react';
 import { EffectEx } from '@dxos/effect';
@@ -17,7 +18,7 @@ import { type ActionGraphProps, Menu, MenuBuilder, useMenuBuilder } from '@dxos/
 
 import { MasterDetail, type MasterDetailAdornment, type MasterDetailIcon, RoutineForm } from '#components';
 import { meta } from '#meta';
-import { Routine, RoutineCapabilities } from '#types';
+import { RoutineCapabilities } from '#types';
 
 import { connectedRoutinesQuery } from '../../util';
 
@@ -108,7 +109,7 @@ export const RoutineCompanion = ({ subject: object, attendableId }: RoutineCompa
       return;
     }
 
-    // The draft is a fully-wired graph (see `Routine.make`); a single add cascades the owned trigger and
+    // The draft is a fully-wired graph (see `makeRoutine`); a single add cascades the owned trigger and
     // instructions, and the subject carried in the instructions' `objects` is the structural connection the
     // query finds.
     const persistedRoutine = db.add(draft);
