+++ packages/plugins/plugin-routine/src/types/index.ts
@@ -4,6 +4,5 @@
 
 export * from './schema';
 
-export * as Routine from './Routine';
 export * as RoutineCapabilities from './RoutineCapabilities';
 export * as RoutineOperation from './RoutineOperation';
