+++ packages/plugins/plugin-routine/src/util/index.ts
@@ -4,3 +4,4 @@
 
 export * from './routines-for-object';
 export * from './run-instructions';
+export * from './wire';
