+++ packages/plugins/plugin-routine/src/util/wire.test.ts
@@ -4,19 +4,18 @@
 
 import { describe, test } from 'vitest';
 
-import { Instructions, Operation, Runnable, Trigger } from '@dxos/compute';
-import { Obj, Ref, Type } from '@dxos/echo';
+import { Instructions, Operation, Routine, Runnable, Trigger } from '@dxos/compute';
+import { Obj, Ref } from '@dxos/echo';
 
 import { blank } from '../templates';
-import { isRunInstructions } from '../util';
-import * as Routine from './Routine';
+import { isRunInstructions } from './run-instructions';
+import { makeRoutine } from './wire';
 
-describe('Routine', () => {
-  test('make produces a typed Routine', ({ expect }) => {
-    const routine = Routine.make({ name: 'Test', triggers: [] });
+describe('wire', () => {
+  test('makeRoutine produces a typed Routine', ({ expect }) => {
+    const routine = makeRoutine({ name: 'Test', triggers: [] });
     expect(Routine.instanceOf(routine)).toBe(true);
     expect(Obj.instanceOf(Routine.Routine, routine)).toBe(true);
-    expect(Type.getTypename(Routine.Routine)).toBe('org.dxos.type.routine');
     expect(Obj.getLabel(routine)).toBe('Test');
     expect(routine.triggers).toEqual([]);
     expect(routine.spec).toBeUndefined();
@@ -27,10 +26,10 @@ describe('Routine', () => {
     expect(Runnable.Runnable).toBe(Operation.PersistentOperation);
   });
 
-  test('make wires an instructions action so a single add yields a runnable routine', ({ expect }) => {
+  test('makeRoutine wires an instructions action so a single add yields a runnable routine', ({ expect }) => {
     const instructions = Instructions.make({ name: 'Body', text: 'do something' });
     const trigger = Trigger.make({ spec: Trigger.specTimer('0 9 * * *') });
-    const routine = Routine.make({ name: 'R', instructions, trigger });
+    const routine = makeRoutine({ name: 'R', instructions, trigger });
 
     // The routine owns the instructions, and the trigger dispatches RunInstructions with those instructions
     // bound into the trigger input so no separate persistence step is needed.
@@ -43,9 +42,9 @@ describe('Routine', () => {
     expect(trigger.remote).toBeUndefined();
   });
 
-  test('make preserves an explicit trigger remote override', ({ expect }) => {
+  test('makeRoutine preserves an explicit trigger remote override', ({ expect }) => {
     const trigger = Trigger.make({ remote: true });
-    const routine = Routine.make({ name: 'R', trigger });
+    const routine = makeRoutine({ name: 'R', trigger });
     expect(routine.triggers[0]?.target?.remote).toBe(true);
   });
